Moved setCmdVel to kinematic_common_interface
